Renaming the files, moving to correct packages
